split out UserControl
