+++ frontend/src/components/core/header/SignOutNavLink.tsx
@@ -1,23 +1,9 @@
-import { testApplicationId } from "src/constants/auth";
 import { useUser } from "src/services/auth/useUser";
 
 import { useTranslations } from "next-intl";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useCallback } from "react";
-
-// links directly to a test application, only used in local environments when logged in as specific test user
-export const TestApplicationLink = () => {
-  const t = useTranslations("Header.navLinks");
-  return (
-    <Link
-      className="display-flex usa-button usa-button--unstyled text-no-underline"
-      href={`/workspace/applications/${testApplicationId}`}
-    >
-      {t("testApplication")}
-    </Link>
-  );
-};
 
 /** Sign out as a nav dropdown child—same structure as NavLink (Link + div) so it matches other menu items */
 export const SignOutNavLink = ({ onClick }: { onClick: () => void }) => {
